requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pyautogui==0.9.54
websockets==12.0
python-multipart==0.0.9
mss==9.0.1
Pillow==10.2.0
pyperclip==1.8.2
pyinstaller==6.4.0
